pyproject.toml
[project]
name = "telegram-channel-automation-bot"
version = "1.0.0"
description = "Private inline-only control panel for one Telegram channel (webhook + PostgreSQL + APScheduler)."
requires-python = ">=3.12"
dependencies = [
  "python-telegram-bot==21.0.1",
  "fastapi==0.109.2",
  "uvicorn[standard]==0.27.1",
  "sqlalchemy[asyncio]==2.0.25",
  "asyncpg==0.29.0",
  "apscheduler==3.10.4",
  "python-dotenv==1.0.1",
  "pydantic-settings==2.1.0",
  "httpx==0.26.0",
  "greenlet==3.0.3",
  "tzdata==2024.1",
]

[project.scripts]
channel-bot = "bot.main:main"
